Separate output string list use new lines: getline
The --separate-output parameter is ignored and spaces are still used as a delimiter instead of a new line character when interpreting the output from the POSIX functions popen() and getline(). This should fix the issue. I don't know how to build this, so I can't verify whether it wiill work. I will provide code for anyone who is of the mind to test it themselves.