Fixes bug 472715: use Qt directory selector to select an existing directory
Unlike what I said in the connected bug report, the problem was not in the removal of KDirSelectDialog but in the incorrect addition of `setFileMode()`.

I went a bit back and forth on the minimal fix (I should have tested more), and I apologize for the noise.

Hopefully you can accept this minimal change quickly.


BUG: 472715